estate-intake-requirements.txt
pydantic>=2.0.0
langchain>=0.2
langchain-openai
python-dotenv
pyyaml
pillow
rich
tqdm
pytest
click
